Main function of adding one coin is complete and portfolio keeps track of current value

+++ src/model/Portfolio.java
@@ -1,17 +1,17 @@
 package model;
 
 public class Portfolio {
-    private int value;
+    private double value;
 
     public Portfolio() {
         value = 0;
     }
 
-    public void setValue(int i) {
+    public void setValue(double i) {
         this.value = i;
     }
 
-    public int getValue() {
+    public double getValue() {
         return value;
     }
 }
